Question 1 TD3 finie

+++ shop.pizza-shop/src/domain/dto/ProduitDTO.php
@@ -10,7 +10,7 @@ class ProduitDTO extends \pizzashop\shop\domain\dto\DTO
     public string $libelle_produit;
     public string $libelle_categorie;
     public string $libelle_taille;
-    public int $tarif;
+    public float $tarif;
 
     /**
      * @param int $numero_produit
